// frontend/generated/models/AccountDto.ts
export interface AccountDto {
    /**
     * 
     * @type {number}
     * @memberof AccountDto
     */
    accountId?: number;
    /**
     * 
     * @type {string}
     * @memberof AccountDto
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof AccountDto
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof AccountDto
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof AccountDto
     */
    password?: string;
    /**
     * 
     * @type {string}
     * @memberof AccountDto
     */
    avatarUrl?: string;
    /**
     * 
     * @type {number}
     * @memberof AccountDto
     */
    balance?: number;
    /**
     * 
     * @type {string}
     * @memberof AccountDto
     */
    role?: AccountDtoRoleEnum;
    /**
     * 
     * @type {boolean}
     * @memberof AccountDto
     */
    isVerified?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof AccountDto
     */
    verifiedAt?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof AccountDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof AccountDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof AccountDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof AccountDto
     */
    profileId?: number;
}


/**
 * @export
 */
export const AccountDtoRoleEnum = {
    Admin: 'ADMIN',
    Staff: 'STAFF',
    Client: 'CLIENT',
    Freelancer: 'FREELANCER'
} as const;
export type AccountDtoRoleEnum = typeof AccountDtoRoleEnum[keyof typeof AccountDtoRoleEnum];