// frontend/generated/models/ContractDto.ts
export interface ContractDto {
    /**
     * 
     * @type {number}
     * @memberof ContractDto
     */
    contractId?: number;
    /**
     * 
     * @type {ProposalDto}
     * @memberof ContractDto
     */
    proposal?: ProposalDto;
    /**
     * 
     * @type {AccountDto}
     * @memberof ContractDto
     */
    freelancer?: AccountDto;
    /**
     * 
     * @type {number}
     * @memberof ContractDto
     */
    projectId?: number;
    /**
     * 
     * @type {number}
     * @memberof ContractDto
     */
    budget?: number;
    /**
     * 
     * @type {Date}
     * @memberof ContractDto
     */
    signedAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ContractDto
     */
    terminatedAt?: Date;
    /**
     * 
     * @type {ContractStatusDto}
     * @memberof ContractDto
     */
    status?: ContractStatusDto;
    /**
     * 
     * @type {FileDto}
     * @memberof ContractDto
     */
    contractFile?: FileDto;
    /**
     * 
     * @type {Date}
     * @memberof ContractDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ContractDto
     */
    updatedAt?: Date;
}