// frontend/generated/models/MessageDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FileDto } from './FileDto';
import {
    FileDtoFromJSON,
    FileDtoFromJSONTyped,
    FileDtoToJSON,
    FileDtoToJSONTyped,
} from './FileDto';
import type { AccountDto } from './AccountDto';
import {
    AccountDtoFromJSON,
    AccountDtoFromJSONTyped,
    AccountDtoToJSON,
    AccountDtoToJSONTyped,
} from './AccountDto';

/**
 * 
 * @export
 * @interface MessageDto
 */
export interface MessageDto {
    /**
     * 
     * @type {number}
     * @memberof MessageDto
     */
    messageId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof MessageDto
     */
    sender?: AccountDto;
    /**
     * 
     * @type {number}
     * @memberof MessageDto
     */
    projectId?: number;
    /**
     * 
     * @type {string}
     * @memberof MessageDto
     */
    content?: string;
    /**
     * 
     * @type {boolean}
     * @memberof MessageDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Array<FileDto>}
     * @memberof MessageDto
     */
    files?: Array<FileDto>;
    /**
     * 
     * @type {Date}
     * @memberof MessageDto
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the MessageDto interface.
 */
export function instanceOfMessageDto(value: object): value is MessageDto {
    return true;
}

export function MessageDtoFromJSON(json: any): MessageDto {
    return MessageDtoFromJSONTyped(json, false);
}

export function MessageDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MessageDto {
    if (json == null) {
        return json;
    }
    return {
        
        'messageId': json['messageId'] == null ? undefined : json['messageId'],
        'sender': json['sender'] == null ? undefined : AccountDtoFromJSON(json['sender']),
        'projectId': json['projectId'] == null ? undefined : json['projectId'],
        'content': json['content'] == null ? undefined : json['content'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'files': json['files'] == null ? undefined : ((json['files'] as Array<any>).map(FileDtoFromJSON)),
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function MessageDtoToJSON(json: any): MessageDto {
    return MessageDtoToJSONTyped(json, false);
}

export function MessageDtoToJSONTyped(value?: MessageDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'messageId': value['messageId'],
        'sender': AccountDtoToJSON(value['sender']),
        'projectId': value['projectId'],
        'content': value['content'],
        'isVisible': value['isVisible'],
        'files': value['files'] == null ? undefined : ((value['files'] as Array<any>).map(FileDtoToJSON)),
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
    };
}