// frontend/generated/models/MilestoneStatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const MilestoneStatusDto = {
    Pending: 'PENDING',
    Terminated: 'TERMINATED',
    InProgress: 'IN_PROGRESS',
    Reviewing: 'REVIEWING',
    Finished: 'FINISHED'
} as const;
export type MilestoneStatusDto = typeof MilestoneStatusDto[keyof typeof MilestoneStatusDto];


export function instanceOfMilestoneStatusDto(value: any): boolean {
    for (const key in MilestoneStatusDto) {
        if (Object.prototype.hasOwnProperty.call(MilestoneStatusDto, key)) {
            if (MilestoneStatusDto[key as keyof typeof MilestoneStatusDto] === value) {
                return true;
            }
        }
    }
    return false;
}

export function MilestoneStatusDtoFromJSON(json: any): MilestoneStatusDto {
    return MilestoneStatusDtoFromJSONTyped(json, false);
}

export function MilestoneStatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MilestoneStatusDto {
    return json as MilestoneStatusDto;
}

export function MilestoneStatusDtoToJSON(value?: MilestoneStatusDto | null): any {
    return value as any;
}

export function MilestoneStatusDtoToJSONTyped(value: any, ignoreDiscriminator: boolean): MilestoneStatusDto {
    return value as MilestoneStatusDto;
}