// frontend/generated/models/ProposalDto.ts
export interface ProposalDto {
    /**
     * 
     * @type {number}
     * @memberof ProposalDto
     */
    proposalId?: number;
    /**
     * 
     * @type {number}
     * @memberof ProposalDto
     */
    projectId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof ProposalDto
     */
    freelancer?: AccountDto;
    /**
     * 
     * @type {string}
     * @memberof ProposalDto
     */
    notes?: string;
    /**
     * 
     * @type {number}
     * @memberof ProposalDto
     */
    budget?: number;
    /**
     * 
     * @type {ProposalStatusDto}
     * @memberof ProposalDto
     */
    status?: ProposalStatusDto;
    /**
     * 
     * @type {number}
     * @memberof ProposalDto
     */
    contractId?: number;
    /**
     * 
     * @type {Date}
     * @memberof ProposalDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ProposalDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {Array<FileDto>}
     * @memberof ProposalDto
     */
    files?: Array<FileDto>;
}